Extract connection recovery package
