+++ flake.nix
@@ -120,6 +120,17 @@
                     ];
                 };
 
+                agentConnectivitySource = nix-filter.lib {
+                    root = ./packages/agent-connectivity;
+                    include = [
+                        "cbits"
+                        "src"
+                        "test"
+                        "agent-connectivity.cabal"
+                        "LICENSE"
+                    ];
+                };
+
                 agentRuntimeDaemonSource = nix-filter.lib {
                     root = ./packages/agent-runtime-daemon;
                     include = [
@@ -492,6 +503,14 @@
                             {
                                 src = agentProcessSource;
                             });
+                        agent-connectivity = localPackage
+                            (pkgs.haskell.lib.overrideSrc
+                                (final.callPackage
+                                    ./packages/agent-connectivity/package.nix
+                                    { })
+                                {
+                                    src = agentConnectivitySource;
+                                });
                         agent-runtime-daemon = localPackage
                             (pkgs.haskell.lib.overrideSrc
                                 (final.callPackage
@@ -640,6 +659,8 @@
                 agentMcpPackage = productionHaskellPackages.agent-mcp;
                 agentJsonPackage = productionHaskellPackages.agent-json;
                 agentProcessPackage = productionHaskellPackages.agent-process;
+                agentConnectivityPackage =
+                    productionHaskellPackages.agent-connectivity;
                 agentRuntimeDaemonPackage =
                     productionHaskellPackages.agent-runtime-daemon;
                 agentCodexDialectPackage = productionHaskellPackages.agent-codex-dialect;
@@ -940,6 +961,7 @@
                 packages.agent-mcp = agentMcpPackage;
                 packages.agent-json = agentJsonPackage;
                 packages.agent-process = agentProcessPackage;
+                packages.agent-connectivity = agentConnectivityPackage;
                 packages.agent-runtime-daemon = agentRuntimeDaemonExecutable;
                 packages.agent-codex-dialect = agentCodexDialectPackage;
                 packages.agent-grok-build-dialect = agentGrokBuildDialectPackage;
@@ -983,6 +1005,7 @@
                         packages.agent-mcp
                         packages.agent-json
                         packages.agent-process
+                        packages.agent-connectivity
                         packages.agent-runtime-daemon
                         packages.agent-codex-dialect
                         packages.agent-grok-build-dialect
@@ -1040,6 +1063,7 @@
                     agent-mcp = haskellPackages.agent-mcp;
                     agent-json = haskellPackages.agent-json;
                     agent-process = haskellPackages.agent-process;
+                    agent-connectivity = haskellPackages.agent-connectivity;
                     agent-runtime-daemon =
                         haskellPackages.agent-runtime-daemon;
                     agent-codex-dialect = haskellPackages.agent-codex-dialect;
