+++ packages/agent-cli/agent-cli.cabal
@@ -57,7 +57,6 @@ library
                     , Agent.CLI.Clipboard.Linux
                     , Agent.CLI.Clipboard.MacOS
                     , Agent.CLI.McpStatus
-                    , Agent.CLI.NetworkPath
                     , Agent.CLI.Input.Approval
                     , Agent.CLI.Login.Types
                     , Agent.CLI.Lsp.Capabilities
@@ -150,7 +149,6 @@ library
                     , Agent.CLI.CodeModeRuntime
                     , Agent.CLI.Compaction
                     , Agent.CLI.Context
-                    , Agent.CLI.Connectivity
                     , Agent.CLI.Command
                     , Agent.CLI.ComputerUse
                     , Agent.CLI.Config
@@ -260,6 +258,7 @@ library
                     , agent-cli-runtime:Agent.CLI.Session.StoreCodec as Agent.CLI.Session.StoreCodec
                     , agent-cli-runtime:Agent.CLI.SessionLock as Agent.CLI.SessionLock
     build-depends:    agent-cli-runtime >= 0.1 && < 0.2
+                    , agent-connectivity >= 0.1 && < 0.2
                     , agent-core >= 0.1 && < 0.2
                     , agent-mcp >= 0.1 && < 0.2
                     , agent-json >= 0.1 && < 0.2
@@ -316,11 +315,6 @@ library
                     , vty-crossplatform >= 0.4 && < 0.5
                     , wai
                     , warp
-    if os(darwin)
-        c-sources:    cbits/NetworkPathMonitor.c
-        cc-options:   -fblocks
-        frameworks:   Network
-
 executable agent-cli
     import:           common-options
     hs-source-dirs:   app
@@ -549,6 +543,7 @@ test-suite agent-cli-test
                     , Agent.CLI.WorktreeSpec
     build-depends:    agent-cli
                     , agent-cli-runtime
+                    , agent-connectivity
                     , agent-runtime-daemon
                     , agent-core
                     , agent-mcp
