+++ packages/agent-cli/package.nix
@@ -1,5 +1,5 @@
 { mkDerivation, aeson, agent-claude, agent-cli-runtime
-, agent-codex-dialect, agent-core, agent-gemini
+, agent-codex-dialect, agent-connectivity, agent-core, agent-gemini
 , agent-grok-build-dialect, agent-json, agent-mcp, agent-openai
 , agent-openrouter, agent-process, agent-responses
 , agent-responses-types, agent-runtime-daemon, agent-store
@@ -22,17 +22,18 @@ mkDerivation {
   # cabal2nix does not emit a separate foreign-library dependency list.
   # Keep agent-runtime-daemon here for the Darwin bridge component.
   libraryHaskellDepends = [
-    aeson agent-claude agent-cli-runtime agent-codex-dialect agent-core
-    agent-gemini agent-grok-build-dialect agent-json agent-mcp
-    agent-openai agent-openrouter agent-process agent-responses
-    agent-runtime-daemon
-    agent-responses-types agent-store agent-syntax agent-tui agent-xai
-    ansi-terminal async base base64-bytestring brick bytestring colour
-    containers crypton directory entropy filelock filepath haskeline
-    hasql-pool http-client http-client-tls http-types JuicyPixels
-    memory mtl network network-uri optparse-applicative process retry
-    safe-exceptions scientific stm tagsoup text time transformers unix
-    vector vty vty-crossplatform wai warp
+    aeson agent-claude agent-cli-runtime agent-codex-dialect
+    agent-connectivity agent-core agent-gemini agent-grok-build-dialect
+    agent-json agent-mcp agent-openai agent-openrouter agent-process
+    agent-responses agent-runtime-daemon
+    agent-responses-types agent-store agent-syntax
+    agent-tui agent-xai ansi-terminal async base base64-bytestring
+    brick bytestring colour containers crypton directory entropy
+    filelock filepath haskeline hasql-pool http-client http-client-tls
+    http-types JuicyPixels memory mtl network network-uri
+    optparse-applicative process retry safe-exceptions scientific stm
+    tagsoup text time transformers unix vector vty vty-crossplatform
+    wai warp
   ];
   executableHaskellDepends = [
     aeson agent-cli-runtime agent-responses agent-responses-types
@@ -40,14 +41,14 @@ mkDerivation {
     safe-exceptions text time unix
   ];
   testHaskellDepends = [
-    aeson agent-claude agent-cli-runtime agent-codex-dialect agent-core
-    agent-gemini agent-grok-build-dialect agent-json agent-mcp
-    agent-openai agent-openrouter agent-responses agent-responses-types
-    agent-runtime-daemon agent-store agent-tui agent-xai ansi-terminal
-    async base brick bytestring colour containers directory filepath
-    haskeline hspec http-client http-types JuicyPixels process
-    QuickCheck safe-exceptions stm text time transformers unix vty wai
-    warp
+    aeson agent-claude agent-cli-runtime agent-codex-dialect
+    agent-connectivity agent-core agent-gemini agent-grok-build-dialect
+    agent-json agent-mcp agent-openai agent-openrouter agent-responses
+    agent-responses-types agent-runtime-daemon agent-store agent-tui
+    agent-xai ansi-terminal async base brick bytestring colour
+    containers directory filepath haskeline hspec http-client
+    http-types JuicyPixels process QuickCheck safe-exceptions stm text
+    time transformers unix vty wai warp
   ];
   benchmarkHaskellDepends = [
     aeson agent-core agent-json agent-responses agent-responses-types
