+++ packages/agent-cli/src/Agent/CLI/Runtime/Repl/Selection.hs
@@ -30,7 +30,7 @@ import Agent.CLI.Config
     ( HarnessConfig(configTheme)
     , updateHarnessConfig
     )
-import Agent.CLI.Connectivity ()
+import Agent.Connectivity ()
 import Agent.CLI.Database ()
 import Agent.CLI.Database.Store ()
 import Agent.CLI.Dialects ()
