+++ packages/agent-connectivity/test/Spec.hs
@@ -1,0 +1,7 @@
+module Main (main) where
+
+import qualified Agent.ConnectivitySpec as ConnectivitySpec
+import Test.Hspec (hspec)
+
+main :: IO ()
+main = hspec ConnectivitySpec.spec
